-REFACTOR: rabbitMq and NotificationModule

+++ common-client-libs/src/main/java/eu/getsoftware/hotelico/clients/infrastructure/notification/NotificationRequest.java
@@ -1,7 +1,7 @@
 package eu.getsoftware.hotelico.clients.infrastructure.notification;
 
 public record NotificationRequest(
-        Integer toCustomerId,
+        Long toCustomerId,
         String toCustomerName,
         String message
 ) {
